Create N-QUEEN.cpp
This code solves the N-Queens problem by recursively trying to place queens on the board, checking if the placement is safe using the isSafe function. If a valid solution is found, it's added to the results vector. If no solution is found, the program backtracks by changing the placement of previous queens and trying a different approach.